tweaking raspberrypi_setup.sh

+++ raspberrypi_setup.sh
@@ -70,9 +70,17 @@ fi
 sudo perl -MCPAN -e 'my $c = "CPAN::HandleConfig"; $c->load(doit => 1, autoconfig => 1); $c->edit(prerequisites_policy => "follow"); $c->edit(build_requires_install_policy => "yes"); $c->commit'
 sudo -H pip  install --upgrade ephem pytz pika python-dateutil
 sudo -H pip3 install --upgrade ephem pytz pika python-dateutil
-sudo -H python2.7 -m pip install --upgrade tendo paho-mqtt smbus-cffi
-sudo -H python3.4 -m pip install --upgrade tendo paho-mqtt smbus-cffi
-sudo -H python3.5 -m pip install --upgrade tendo paho-mqtt smbus-cffi
-sudo -H python3.6 -m pip install --upgrade tendo paho-mqtt smbus-cffi
+if [ -f `which python2.7` ]; then
+    sudo -H python2.7 -m pip install --upgrade tendo paho-mqtt smbus-cffi
+fi
+if [ -f `which python3.4` ]; then
+    sudo -H python3.4 -m pip install --upgrade tendo paho-mqtt smbus-cffi
+fi
+if [ -f `which python3.5` ]; then
+    sudo -H python3.5 -m pip install --upgrade tendo paho-mqtt smbus-cffi
+fi
+if [ -f `which python3.6` ]; then
+    sudo -H python3.6 -m pip install --upgrade tendo paho-mqtt smbus-cffi
+fi
 
 sudo apt-get clean && sudo apt-get autoremove
